Refactor Nat Services
