+++ src/core.rs
@@ -7,6 +7,7 @@ use std::fmt;
 
 use self::flexible_gpus::FlexibleGpu;
 use self::load_balancers::LoadBalancer;
+use self::nat_services::NatServices;
 use self::oos::Oos;
 use self::public_ips::PublicIp;
 use self::snapshots::Snapshot;
@@ -18,6 +19,7 @@ static HOURS_PER_MONTH: f32 = (365_f32 * 24_f32) / 12_f32;
 
 pub mod flexible_gpus;
 pub mod load_balancers;
+pub mod nat_services;
 pub mod oos;
 pub mod public_ips;
 pub mod snapshots;
@@ -181,36 +183,6 @@ impl error::Error for ResourceError {}
 trait ResourceTrait {
     fn price_per_hour(&self) -> Result<f32, ResourceError>;
     fn compute(&mut self) -> Result<(), ResourceError>;
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct NatServices {
-    pub osc_cost_version: Option<String>,
-    pub account_id: Option<String>,
-    pub read_date_rfc3339: Option<String>,
-    pub region: Option<String>,
-    pub resource_id: Option<String>,
-    pub price_product_per_nat_service_per_hour: Option<f32>,
-    pub price_per_hour: Option<f32>,
-    pub price_per_month: Option<f32>,
-}
-
-impl ResourceTrait for NatServices {
-    fn price_per_hour(&self) -> Result<f32, ResourceError> {
-        match self.price_per_hour {
-            Some(price) => Ok(price),
-            None => Err(ResourceError::NotComputed),
-        }
-    }
-    fn compute(&mut self) -> Result<(), ResourceError> {
-        let mut price_per_hour = 0_f32;
-        if let Some(price_non_attached) = self.price_product_per_nat_service_per_hour {
-            price_per_hour += price_non_attached;
-        }
-        self.price_per_hour = Some(price_per_hour);
-        self.price_per_month = Some(price_per_hour * HOURS_PER_MONTH);
-        Ok(())
-    }
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
