+++ src/oapi.rs
@@ -1,5 +1,5 @@
 use crate::args::Filter;
-use crate::core::{self, Resources};
+use crate::core::{self};
 use crate::VERSION;
 use aws_config::SdkConfig;
 use aws_credential_types::provider::SharedCredentialsProvider;
@@ -12,14 +12,12 @@ use outscale_api::apis::catalog_api::read_catalog;
 use outscale_api::apis::configuration::AWSv4Key;
 use outscale_api::apis::configuration::Configuration;
 use outscale_api::apis::configuration_file::{ConfigurationFile, ConfigurationFileError};
-use outscale_api::apis::nat_service_api::read_nat_services;
 use outscale_api::apis::subregion_api::read_subregions;
 use outscale_api::apis::Error::ResponseError;
 use outscale_api::models::{
-    Account, CatalogEntry, FiltersNatService, FlexibleGpu, Image, NatService, PublicIp,
-    ReadAccountsRequest, ReadAccountsResponse, ReadCatalogRequest, ReadCatalogResponse,
-    ReadNatServicesRequest, ReadNatServicesResponse, ReadSubregionsRequest, ReadSubregionsResponse,
-    Snapshot, Vm, VmType, Volume,
+    Account, CatalogEntry, FlexibleGpu, Image, NatService, PublicIp, ReadAccountsRequest,
+    ReadAccountsResponse, ReadCatalogRequest, ReadCatalogResponse, ReadSubregionsRequest,
+    ReadSubregionsResponse, Snapshot, Vm, VmType, Volume,
 };
 use rand::rngs::ThreadRng;
 use rand::{thread_rng, Rng};
@@ -34,6 +32,7 @@ use std::time::Duration;
 
 use self::flexible_gpus::FlexibleGpuId;
 use self::load_balancers::LoadbalancerId;
+use self::nat_services::NatServiceId;
 use self::oos::{BucketId, OosBucket};
 use self::public_ips::PublicIpId;
 use self::snapshots::SnapshotId;
@@ -45,7 +44,6 @@ static THROTTLING_MIN_WAIT_MS: u64 = 1000;
 static THROTTLING_MAX_WAIT_MS: u64 = 10000;
 
 type ImageId = String;
-type NatServiceId = String;
 
 // This string correspond to pure internal forged identifier of a catalog entry.
 // format!("{}/{}/{}", entry.service, entry.type, entry.operation);
@@ -55,6 +53,7 @@ type VmTypeName = String;
 
 mod flexible_gpus;
 mod load_balancers;
+mod nat_services;
 mod oos;
 mod public_ips;
 mod snapshots;
@@ -312,47 +311,6 @@ impl Input {
         Ok(())
     }
 
-    fn fetch_nat_services(&mut self) -> Result<(), Box<dyn error::Error>> {
-        let result: ReadNatServicesResponse = loop {
-            let filters: FiltersNatService = match &self.filters {
-                Some(filter) => FiltersNatService {
-                    tag_keys: Some(filter.filter_tag_key.clone()),
-                    tag_values: Some(filter.filter_tag_value.clone()),
-                    tags: Some(filter.filter_tag.clone()),
-                    ..Default::default()
-                },
-                None => FiltersNatService::new(),
-            };
-            let request = ReadNatServicesRequest {
-                filters: Some(Box::new(filters)),
-                ..Default::default()
-            };
-            let response = read_nat_services(&self.config, Some(request));
-            if Input::is_throttled(&response) {
-                self.random_wait();
-                continue;
-            }
-            break response?;
-        };
-        let nat_services = match result.nat_services {
-            None => {
-                warn!("no nat_service available!");
-                return Ok(());
-            }
-            Some(nat_services) => nat_services,
-        };
-
-        for nat_service in nat_services {
-            let nat_service_id = nat_service
-                .nat_service_id
-                .clone()
-                .unwrap_or_else(|| String::from(""));
-            self.nat_services.insert(nat_service_id, nat_service);
-        }
-        info!("fetched {} nat_service", self.nat_services.len());
-        Ok(())
-    }
-
     fn random_wait(&mut self) {
         let wait_time_ms = self
             .rng
@@ -395,30 +353,6 @@ impl Input {
                 warn!("cannot find catalog entry for {}", entry_id);
                 None
             }
-        }
-    }
-
-    fn fill_resource_nat_service(&self, resources: &mut Resources) {
-        for (nat_service_id, nat_service) in &self.nat_services {
-            let price_product_per_nat_service_per_hour =
-                self.catalog_entry("TinaOS-FCU", "NatGatewayUsage", "CreateNatGateway");
-            let Some(nat_service_id) = &nat_service.nat_service_id else {
-                warn!("cannot get nat_service_id content for {}", nat_service_id );
-                continue;
-            };
-            let core_nat_service = core::NatServices {
-                osc_cost_version: Some(String::from(VERSION)),
-                account_id: self.account_id(),
-                read_date_rfc3339: self.fetch_date.map(|date| date.to_rfc3339()),
-                region: self.region.clone(),
-                resource_id: Some(nat_service_id.clone()),
-                price_per_hour: None,
-                price_per_month: None,
-                price_product_per_nat_service_per_hour,
-            };
-            resources
-                .resources
-                .push(core::Resource::NatServices(core_nat_service));
         }
     }
 }
